src/backend/pyproject.toml
[project]
name = "mtl-finder-backend"
version = "0.1.0"
description = "FastAPI backend for the MTL Finder application"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.0",
    "gtfs-realtime-bindings>=2.0.0",
    "mistralai>=1.10.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.39.0",
]
